requirements.txt.txt
pyTelegramBotAPI==4.13.0
openai==1.0.0
pillow==9.5.0
pytesseract==0.3.10
reportlab==3.6.12
beautifulsoup4==4.12.2
requests==2.31.0